+++ examples/magento-graphcms/components/Layout/LayoutFull.tsx
@@ -38,7 +38,7 @@ export function LayoutFull(props: LayoutFullProps) {
   return (
     <LayoutDefault
       {...uiProps}
-      noSticky={router.asPath === '/'}
+      noSticky={router.asPath.split('?')[0] === '/'}
       header={
         <>
           <Logo />