+++ packages/next-ui/LayoutParts/GlobalHead.tsx
@@ -25,9 +25,10 @@ export function GlobalHead(props: GlobalHeadProps) {
       <meta name='apple-mobile-web-app-title' content={name} key='apple-mobile-web-app-title' />
       <meta name='format-detection' content='telephone=no' key='format-detection' />
       <meta name='mobile-web-app-capable' content='yes' key='mobile-web-app-capable' />
-      <link rel='apple-touch-icon' href='/manifest/icon-512-512.png' key='apple-touch-icon' />
+      <link rel='icon' href='/favicon.ico' sizes='any' />
+      <link rel='icon' href='/favicon.svg' type='image/svg+xml' />
+      <link rel='apple-touch-icon' href='/apple-touch-icon.png' />
       <link rel='manifest' href='/manifest.webmanifest' key='manifest' />
-      <link rel='icon' href='/manifest/favicon.ico' key='icon' />
     </Head>
   )
 }